refactored some routes
